[WIP] rewriting terrain system

+++ GameEngine2/src/main/java/EngineTester/MainGameLoop.java
@@ -109,7 +109,7 @@ public class MainGameLoop {
         WaterShader waterShader = new WaterShader();
         WaterRenderer waterRenderer = new WaterRenderer(loader, waterShader, masterRenderer.getProjectionMatrix(), waterFrameBuffers);
         List<WaterTile> waters = new ArrayList<WaterTile>();
-        WaterTile mainWaterTile = new WaterTile((float) (TERRAIN_SIZE/2.0), (float) (TERRAIN_SIZE/2.0), 0, TERRAIN_SIZE/2);
+        WaterTile mainWaterTile = new WaterTile((float) (TERRAIN_SIZE/2.0), (float) (TERRAIN_SIZE/2.0), 0, (float) (TERRAIN_SIZE/2.0));
         waters.add(mainWaterTile);
 
         //Game loop
